+++ redline/api/src/main/java/run/endive/redline/experimental/api/Interruptible.java
@@ -1,4 +1,4 @@
-package run.endive.redline.api;
+package run.endive.redline.experimental.api;
 
 public interface Interruptible {
     void requestInterrupt();
