+++ redline/api/src/main/java/run/endive/redline/experimental/api/RedlineTarget.java
@@ -1,4 +1,4 @@
-package run.endive.redline.api;
+package run.endive.redline.experimental.api;
 
 import java.util.Locale;
 import java.util.Optional;
