+++ redline/api/src/main/java/run/endive/redline/experimental/api/internal/CtxBuffer.java
@@ -1,4 +1,4 @@
-package run.endive.redline.api.internal;
+package run.endive.redline.experimental.api.internal;
 
 /**
  * Layout of the shared context buffer passed between Java and native compiled code.
