+++ redline/compiler/src/main/java/run/endive/redline/experimental/compiler/internal/NativeEmitters.java
@@ -1,6 +1,6 @@
-package run.endive.redline.compiler.internal;
+package run.endive.redline.experimental.compiler.internal;
 
-import run.endive.redline.api.internal.CtxBuffer;
+import run.endive.redline.experimental.api.internal.CtxBuffer;
 import run.endive.wasm.types.AnnotatedInstruction;
 import run.endive.wasm.types.ExternalType;
 import run.endive.wasm.types.FunctionType;
