+++ redline/runner/src/main/java/run/endive/redline/experimental/runner/internal/NativeGlobalInstance.java
@@ -1,4 +1,4 @@
-package run.endive.redline.runner.internal;
+package run.endive.redline.experimental.runner.internal;
 
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
