+++ redline/runner/src/main/java/run/endive/redline/experimental/runner/internal/NativeMemory.java
@@ -1,4 +1,4 @@
-package run.endive.redline.runner.internal;
+package run.endive.redline.experimental.runner.internal;
 
 import static run.endive.runtime.ConstantEvaluators.computeConstantValue;
