+++ redline/runner/src/main/java/run/endive/redline/experimental/runner/internal/NativeTable.java
@@ -1,11 +1,11 @@
-package run.endive.redline.runner.internal;
+package run.endive.redline.experimental.runner.internal;
 
 import static run.endive.wasm.types.Value.REF_NULL_VALUE;
 
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
-import run.endive.redline.api.internal.CtxBuffer;
+import run.endive.redline.experimental.api.internal.CtxBuffer;
 import run.endive.runtime.Instance;
 import run.endive.runtime.TableInstance;
 import run.endive.wasm.WasmEngineException;
